% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGADownload.R
\name{TCGADownload}
\alias{TCGADownload}
\title{TCGA Download}
\usage{
TCGADownload(fileURL, earlyStop = 0)
}
\arguments{
\item{fileURL}{The character list of URLs referring to the exact location of the data queried}

\item{earlyStop}{If the number of file is too large for you need you can choose a subset}
}
\description{
Download data previously selected using the TCGAQuery function
TCGADownload(fileURL, earlyStop = 0)
}
\author{
Davide
}
\seealso{
TCGAQuery
}

