\name{TCGADownloader-package}
\alias{TCGADownloader-package}
\alias{TCGADownloader}
\docType{package}
\title{
What the package does (short line)
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab TCGADownloader\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-18\cr
License: \tab Davide\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Davide

Maintainer: Who to complain to <yourfault@somewhere.net>

}


\keyword{ package }