% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGAQuery.R
\name{TCGAQuery}
\alias{TCGAQuery}
\title{TCGA query}
\usage{
TCGAQuery(tumor = "all", centerType = "all", center = "all",
  platform = "all", level = "all", version = "all", i = F,
  file = "inst/extdata/dataMatrix.rda")
}
\arguments{
\item{tumor}{tumor code}

\item{centerType}{type code}

\item{center}{center code}

\item{platform}{platform code}

\item{level}{level 1 2 3}

\item{version}{version code -TO DO-}

\item{i}{- interactive -TO DO-}

\item{file}{- link reference data matrix}
}
\description{
Crossfinding of file locations for downloading (TCGADownload)
TCGAQuery(tumor = "all",centerType = "all",center = "all",
platform = "all",level = "all",version = "all",i = F,file = "inst/extdata/dataMatrix.rda")
}
\author{
Davide
}
\seealso{
TCGADownload
}

