% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGAUpdate.R
\name{TCGAUpdate}
\alias{TCGAUpdate}
\title{TCGAUpdate}
\usage{

  TCGAUpdate(siteTCGA = "https://tcga-data.nci.nih.gov/tcgafiles/ftp_auth/distro_ftpusers/anonymous/tumor/",
  tableLocation = "data/")
}
\arguments{
\item{siteTCGA}{The starting ftp URI}

\item{tableLocation}{The place in which you keep the data matrix}
}
\description{
replace the dataFolders.rda matrix with the new matrix from "The Cancer Genome Atlas" ftp
TCGAUpdate(
siteTCGA
,tableLocation="data/")
}
\author{
Davide
}
\seealso{
TCGAQuery
}

