% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGADownload.R
\name{TCGADownload}
\alias{TCGADownload}
\title{TCGA Download}
\usage{
TCGADownload(dirURL = "data/query/", finalDir = "data/final",
  earlyStop = 0)
}
\arguments{
\item{dirURL}{the location of the query saved URL links.}

\item{finalDir}{location of the final data saving}

\item{earlyStop}{If the number of file is too large for you need you can choose a subset}
}
\description{
Download data previously selected using the TCGAQuery function
TCGADownload(dirURL="data/query/", finalDir = "data/final", earlyStop = 0)
}
\author{
Davide
}
\seealso{
TCGAQuery
}

